refactor: PostTable로 분리
