+++ src/widgets/post/index.ts
@@ -2,3 +2,4 @@ export * from "./ui/PostAddDialog"
 export * from "./ui/PostDetailDialog"
 export * from "./ui/PostEditDialog"
 export * from "./ui/PostFilterBar"
+export * from "./ui/PostTable"
